bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "aiogram-dialog>=2.4.0",
    "babel>=2.17.0",
    "matplotlib>=3.10.8",
    "pillow>=12.1.0",
    "postgres[binary,pool]>=4.0",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "werkzeug>=3.1.5",
]
